requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
sentence-transformers==2.7.0
huggingface_hub==0.20.3
transformers==4.36.2
torch==2.1.2
faiss-cpu==1.7.4
google-generativeai==0.3.2
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
aiofiles==23.2.0
httpx==0.25.2